% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{capitalize_firstLetter}
\alias{capitalize_firstLetter}
\title{Internal function to capitalize the first letter of a character}
\usage{
capitalize_firstLetter(char)
}
\arguments{
\item{char}{Character value whose first letter should be capitalized}
}
\description{
Internal helper function to capitalize the first letter of a character value.
The use case is to create a plot label like 'Age' from a variable name like
'age'.
}
