% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APChexamap_helpers.R
\name{compute_xCoordinate}
\alias{compute_xCoordinate}
\title{Internal helper to tilt the x-axis for the hexamap plot}
\usage{
compute_xCoordinate(period_vec)
}
\arguments{
\item{period_vec}{Numeric vector of period values.}
}
\description{
Internal helper function to be called in \code{\link{plot_APChexamap}},
to tilt the x-axis for the hexamap plot.
}
