% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APChexamap_helpers.R
\name{compute_yCoordinate}
\alias{compute_yCoordinate}
\title{Internal helper to tilt the x-axis for the hexamap plot}
\usage{
compute_yCoordinate(period_vec, age_vec)
}
\arguments{
\item{period_vec}{Numeric vector of period values.}

\item{age_vec}{Numeric vector of age values.}
}
\description{
Internal helper function to be called in \code{\link{plot_APChexamap}},
to tilt the x-axis for the hexamap plot.
}
