% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_APCsummary.R
\name{create_APCsummary}
\alias{create_APCsummary}
\title{Create a summary table for multiple estimated GAM models}
\usage{
create_APCsummary(model_list, dat, digits = 2, ...)
}
\arguments{
\item{model_list}{A list of regression models estimated with
\code{\link[mgcv]{gam}}. If the list is named, the names are used as labels.}

\item{dat}{Dataset as passed to \code{gam} for model estimation, containing
columns \code{period} and \code{age}.}

\item{digits}{Number of digits for numeric columns. Defaults to 2.}

\item{...}{Optional additional arguments passed to \code{\link[knitr]{kable}}.}
}
\description{
Create a table to summarize the overall effect strengths of the age, period
and cohort effects. The output format can be adjusted by passing arguments
to \code{\link[knitr]{kable}} via the \code{...} argument.
}
\details{
If the model was estimated with a log or logit link, the function
automatically performs an exponential transformation of the effect.
}
