% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_helpers.R
\name{create_highlightDiagonalData}
\alias{create_highlightDiagonalData}
\title{Internal helper to create a dataset for ggplot2 to highlight diagonals}
\usage{
create_highlightDiagonalData(dat, highlight_diagonals)
}
\arguments{
\item{dat}{Dataset with columns \code{period} and \code{age} and the
main variable specified through argument \code{y_var}.}

\item{highlight_diagonals}{Optional internal parameter which is only
specified when \code{plot_density} is called from within
\code{plot_densityMatrix}. See \code{\link{plot_densityMatrix}} for details.}
}
\description{
Internal helper function to create a dataset for \code{ggplot2} that can
be used to highlight specific diagonals in a density matrix.
}
