% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modelSummary.R
\name{create_modelSummary}
\alias{create_modelSummary}
\title{Create model summary tables for multiple estimated GAM models}
\usage{
create_modelSummary(model_list, digits = 2, ...)
}
\arguments{
\item{model_list}{list of APC models}

\item{digits}{number of displayed digits}

\item{...}{additional arguments to \code{\link[mgcv]{gam}}}
}
\value{
List of tables created with \code{\link[knitr]{kable}}.
}
\description{
Create publication-ready summary tables of all linear and nonlinear effects
for models fitted with \code{\link[mgcv]{gam}}. The output format of the
tables can be adjusted by passing arguments to \code{\link[knitr]{kable}} via
the \code{...} argument.
}
\details{
If the model was estimated with a log or logit link, the function
automatically performs an exponential transformation of the effects.
}
\examples{
library(APCtools)
library(mgcv)

data(travel)
model <- gam(mainTrip_distance ~ te(age, period) + residence_region +
             household_size + s(household_income), data = travel)

create_modelsummary(list(model), dat = travel)

}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
