% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_deaths}
\alias{drug_deaths}
\title{Drug deaths of white men in the United States}
\format{
A dataframe containing
\describe{
\item{period}{TODO.}
\item{age}{TODO.}
\item{deaths}{TODO.}
\item{population}{TODO.}
\item{death_rate}{TODO.}
}
}
\usage{
data(data_RA)
}
\description{
TODO detailed description, including how to retrieve the data from the online portal
}
\references{
TODO format both references
Jalal, H., & Burke, D. S. (2020). Hexamaps for Age-Period-Cohort Data Visualization and Implementation in R. Epidemiology (Cambridge, Mass.), 31(6), e47.

Centers for Disease Control and Prevention, National Center for Health Statistics. Underlying Cause of Death 1999-2019 on CDC WONDER Online Database, released in 2020. Data are from the Multiple Cause of Death Files, 1999-2019, as compiled from data provided by the 57 vital statistics jurisdictions through the Vital Statistics Cooperative Program. Accessed at http://wonder.cdc.gov/ucd-icd10.html on Dec 6, 2021 6:03:26 AM
}
\keyword{datasets}
