% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APCheatmap.R
\name{ensure_segmentsInPlotRange}
\alias{ensure_segmentsInPlotRange}
\title{Internal helper for gg_addReferenceLines to keep diagonal lines in the plot range}
\usage{
ensure_segmentsInPlotRange(dat_segments, plot_dat)
}
\arguments{
\item{dat_segments}{Dataset containing information on the diagonal reference
lines.}

\item{plot_dat}{Dataset used for creating the heatmap.}
}
\description{
Internal helper function to be called from within
\code{\link{gg_addReferenceLines}}. This function takes the dataset prepared
for adding diagonal reference lines in the plot, checks if some diagonals
exceed the plot limits, cuts them accordingly, if necessary, and again
returns the corrected dataset.
}
