% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linearEffects.R
\name{extract_modelSummary}
\alias{extract_modelSummary}
\title{Extract summary of linear effects in a gam model}
\usage{
extract_modelSummary(model)
}
\arguments{
\item{model}{Model fitted with \code{\link[mgcv]{gam}}.}
}
\description{
Creates a \code{data.frame} containing the linear effects summary of a
model fitted with \code{\link[mgcv]{gam}}.
}
\details{
If the model was estimated with a log link, the function automatically
performs an exponential transformation of the effects.
}
