% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCmodel.R
\name{fit_APC_model}
\alias{fit_APC_model}
\title{Fit APC model without bivariate tensorproduct spline}
\usage{
fit_APC_model(APC_formula, data, ...)
}
\arguments{
\item{APC_formula}{model formula}

\item{data}{data}

\item{...}{}
}
\description{
Fit APC model without bivariate tensorproduct spline
}
