% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smoothEffects.R
\name{get_plotGAMobject}
\alias{get_plotGAMobject}
\title{Extract returned values of plot.gam() while suppressing creation of the plot}
\usage{
get_plotGAMobject(model)
}
\arguments{
\item{model}{GAM model fitted with \code{\link[mgcv]{gam}} or
\code{\link[mgcv]{bam}}.}
}
\description{
Internal helper function to extract the values returned of
\code{\link[mgcv]{plot.gam}} while suppressing creation of the plot.
}
