% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_helpers.R
\name{gg_highlightDiagonals}
\alias{gg_highlightDiagonals}
\title{Internal helper to add the diagonal highlighting to a ggplot}
\usage{
gg_highlightDiagonals(gg, dat, dat_highlightDiagonals)
}
\arguments{
\item{gg}{Existing ggplot object to which the diagonal highlighting should
be added.}

\item{dat}{Dataset with columns \code{period} and \code{age} and the
main variable specified through argument \code{y_var}.}

\item{dat_highlightDiagonals}{Dataset created by
\code{\link{create_highlightDiagonalData}} to highlight specific diagonals
in a density matrix.}
}
\description{
Internal helper function to highlight diagonals in a density matrix. The
function takes an existing ggplot object, adds the diagonal highlighting
and returns the edited ggplot object again.
}
