% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APChexamap.R
\name{plot_APChexamap}
\alias{plot_APChexamap}
\title{Heatmap of the APC surface based on an estimated GAM model}
\usage{
plot_APChexamap(
  data,
  first_age,
  first_period,
  interval,
  first_age_isoline = NULL,
  first_period_isoline = NULL,
  isoline_interval = NULL,
  color_scale = NULL,
  color_map = NULL,
  line_width = 0.5,
  line_color = "grey",
  label_size = 0.5,
  label_color = "black",
  scale_units = "Rate",
  wrap_cohort_labels = TRUE
)
}
\arguments{
\item{wrap_cohort_labels}{}
}
\description{
TEXT
}
