% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density_categorical}
\alias{plot_density_categorical}
\title{Internal helper to plot a categorical density}
\usage{
plot_density_categorical(
  dat,
  y_var,
  weights_var = NULL,
  xlab = y_var,
  ylab = "Density"
)
}
\arguments{
\item{dat}{Dataset with columns \code{period} and \code{age} and the
main variable specified through argument \code{yvar}.}

\item{y_var}{Character name of the main variable to be plotted.}

\item{weights_var}{Optional character name of a weights variable used to
project the results in the sample to some population.}

\item{xlab}{Optional plot annotations.}

\item{ylab}{Optional plot annotations.}
}
\description{
Internal helper function to plot one categorical density, to be called from
within \code{\link{plot_density}}.
}
