% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densityMatrix.R
\name{plot_densityMatrix}
\alias{plot_densityMatrix}
\title{Create a matrix of density plots}
\usage{
plot_densityMatrix(
  dat,
  y_var,
  dimensions = c("period", "age"),
  age_groups = NULL,
  period_groups = NULL,
  cohort_groups = NULL,
  plot_type = "density",
  highlight_diagonals = NULL,
  y_var_cat_breaks = NULL,
  y_var_cat_labels = NULL,
  weights_var = NULL,
  log_scale = FALSE,
  legend_title = NULL,
  ...
)
}
\arguments{
\item{dat}{Dataset with columns \code{period} and \code{age} and the
main variable specified through argument \code{y_var}.}

\item{y_var}{Character name of the main variable to be plotted.}

\item{dimensions}{Character vector specifying the two APC dimensions that
should be visualized along the x-axis and y-axis. Defaults to
\code{c("period","age")}.}

\item{age_groups, period_groups, cohort_groups}{Each a list with one element
specifying the borders of one row or column in the density matrix. E.g., if
the period should be visualized in decade columns from 1980 to 2009, specify
\code{period_groups = list(c(1980,1989), c(1990,1999), c(2000,2009))}.
The list can be named to specify labels for the categories. Only the two
arguments should be passed that were specified by the \code{dimensions}
argument.}

\item{plot_type}{One of \code{c("density","boxplot")}. Only used if the
\code{y_var} column is metric.}

\item{highlight_diagonals}{Optional list to define diagonals in the density
that should be highlighted with different colors. Each list element should be
a numeric vector stating the index of the diagonals (counted from the top
left) that should be highlighted in the same color. If the list is named, the
names are used as legend labels.}

\item{y_var_cat_breaks}{Optional numeric vector of breaks to categorize
\code{y_var} based on calling function \code{\link{cut}}. Only used to
highlight the categories based on different colors. And only used if the
\code{y_var} column is numeric.}

\item{y_var_cat_labels}{Optional character vector for the names of the
categories that were defined based on \code{y_var_cat_breaks}. The length of
this vector must be one shorter than \code{length(y_var_cat_breaks)}. Only
used if the \code{y_var} column is numeric.}

\item{weights_var}{Optional character name of a weights variable used to
project the results in the sample to some population.}

\item{log_scale}{Indicator if the main variable should be log10 transformed.
Only used if the \code{y_var} column is numeric. Defaults to FALSE.}

\item{legend_title}{Optional plot annotations.}

\item{...}{Additional arguments passed to \code{\link{plot_density}}.}
}
\description{
This function creates a matrix of individual density plots or boxplots (for
metric variables) or of individual barplots (for categorical variables).
The age, period or cohort information can each either be plotted on the
x-axis or the y-axis.
}
