% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointMarginalAPCEffects.R
\name{plot_jointMarginalAPCEffects}
\alias{plot_jointMarginalAPCEffects}
\title{Joint plot to compare the marginal APC effects of two models}
\usage{
plot_jointMarginalAPCEffects(
  model1,
  model2,
  dat,
  model_labels = c("model 1", "model 2")
)
}
\arguments{
\item{model1, model2}{Two regression models estimated with
\code{\link[mgcv]{gam}}.}

\item{dat}{Dataset as passed to \code{gam} for model estimation, containing
columns \code{period} and \code{age}.}

\item{model_labels}{Character vector containing the labels for the two models.}
}
\description{
This function creates a joint plot of the marginal APC effects of two
estimated models. It creates a plot with one pane per age, period and
cohort effect, each containing two lines for the two estimated models.
}
\details{
If the model was estimated with a log link, the function automatically
performs an exponential transformation of the effect.
}
