% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partialAPCeffects.R
\name{plot_marginalAPCeffects}
\alias{plot_marginalAPCeffects}
\title{Plot of marginal APC effects based on an estimated GAM model}
\usage{
plot_marginalAPCeffects(model, dat, variable = "age", return_plotData = FALSE)
}
\arguments{
\item{model}{Regression model estimated with \code{\link[mgcv]{gam}}.}

\item{dat}{Dataset as passed to \code{gam} for model estimation, containing
columns \code{period} and \code{age}.}

\item{variable}{One of \code{c("age","period","cohort")}, specifying the
temporal dimension for which the partial effect plots should be created.}

\item{return_plotData}{If TRUE, a list of the datasets prepared for plotting
is returned instead of the ggplot object. The list contains one dataset each
for the overall effect (= evaluations of the APC surface to plot the partial
effects) and for each marginal APC effect (no matter the specified value of
the argument \code{variable}). Defaults to FALSE.}
}
\description{
Plot the marginal effect of age, period or cohort, based on an APC model
estimated as a semiparametric additive regression model with \code{\link[mgcv]{gam}}.
This function is a simple wrapper to \code{\link{plot_partialAPCeffects}},
called with argument \code{hide_partialEffects = TRUE}.
}
