% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridgeline.R
\name{plot_ridgeline}
\alias{plot_ridgeline}
\title{Create a ridgeline plot}
\usage{
plot_ridgeline(
  dat,
  y_var,
  apc_range = NULL,
  y_var_breaks = NULL,
  weights_var = NULL,
  multiple = FALSE,
  xlab = y_var,
  ylab = "Density",
  legend_title = NULL,
  ...
)
}
\arguments{
\item{dat}{Dataset with columns \code{period} and \code{age} and the
main variable specified through argument \code{yvar}.}

\item{y_var}{Character name of the main variable to be plotted.}

\item{apc_range}{Optional list with one or multiple elements with names
\code{"age","period","cohort"} to filter the data. Each element should
contain a numeric vector of values for the respective variable that should
be kept in the data. All other values are deleted.}

\item{y_var_breaks}{Optional numeric vector of breaks to categorize
\code{y_var} based on calling function \code{\link{cut}}. Only used to
highlight the categories based on different colors. And only used if the
\code{y_var} column is numeric.}

\item{weights_var}{Optional character name of a weights variable used to
project the results in the sample to some population.}

\item{multiple}{TODO}

\item{xlab, ylab, legend_title}{Optional plot annotations.}

\item{...}{Additional arguments passed to \code{\link[stats]{density}}.}
}
\description{
Create a ridgeline plot for a specific subset of the data.
}
