% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridgelineMatrix.R
\name{plot_ridgelineMatrix}
\alias{plot_ridgelineMatrix}
\title{Create a matrix of ridgeline plots}
\usage{
plot_ridgelineMatrix(
  data,
  ages = NULL,
  periods = NULL,
  cohorts = NULL,
  log = TRUE,
  adjust = 1
)
}
\description{
Create a matrix of ridgeline plots
}
