% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_variable}
\alias{plot_variable}
\title{Distribution plot of one variable against one APC dimension}
\usage{
plot_variable(dat, variable, apc_dimension = "period")
}
\arguments{
\item{dat}{Dataset containing columns \code{age}, \code{period} and
\code{cohort}.}

\item{variable}{Character name of the variable to plot.}

\item{apc_dimension}{One of \code{c("age","period","cohort")}. Defaults to
\code{"period"}.}
}
\description{
Plot the distribution of one variable in the data against age, period or
cohort. Creates a stacked bar plot for categorical variables and boxplots
for metric variables.
}
