% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariateSummary.R
\name{extract_summary_linearEffects}
\alias{extract_summary_linearEffects}
\title{Internal helper to extract summary of linear effects in a gam model}
\usage{
extract_summary_linearEffects(model)
}
\arguments{
\item{model}{Model fitted with \code{\link[mgcv]{gam}}.}
}
\description{
Internal helper function to create a \code{data.frame} containing the linear
effects summary of a model fitted with \code{\link[mgcv]{gam}}.
}
\details{
If the model was estimated with a log or logit link, the function
automatically performs an exponential transformation of the effect.
}
