% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APCheatmap.R
\name{plot_APCheatmap}
\alias{plot_APCheatmap}
\title{Heatmap of the APC surface based on an estimated GAM model}
\usage{
plot_APCheatmap(model, dat, angle = 15, plot_CI = TRUE)
}
\arguments{
\item{model}{Regression model estimated with \code{\link[mgcv]{gam}}.}

\item{dat}{Dataset as passed to \code{gam} for model estimation, containing
columns \code{period} and \code{age}.}

\item{angle}{Numeric angle of the text labels written on the diagonals that
highlight individual cohorts. Defaults to 15.}

\item{plot_CI}{Indicator if the confidence intervals should be plotted.
Defaults to TRUE.}
}
\description{
Plot the heatmap of the two-dimensional tensor product surface for the
APC effect, based on a model estimated with \code{\link[mgcv]{gam}}.
}
\details{
If the model was estimated with a log link, the function automatically
performs an exponential transformation of the effect.
}
