% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APChexamap.R
\name{plot_APChexamap}
\alias{plot_APChexamap}
\title{Heatmap of the APC surface based on an estimated GAM model}
\usage{
plot_APChexamap(
  data,
  first_age,
  first_period,
  interval,
  first_age_isoline = NULL,
  first_period_isoline = NULL,
  isoline_interval = NULL,
  color_scale = NULL,
  color_map = NULL,
  line_width = 0.5,
  line_color = "grey",
  label_size = 0.5,
  label_color = "black",
  scale_units = "Rate",
  wrap_cohort_labels = TRUE
)
}
\arguments{
\item{data}{TODO}

\item{first_age}{TODO}

\item{first_period}{TODO}

\item{interval}{TODO}

\item{first_age_isoline}{TODO}

\item{first_period_isoline}{TODO}

\item{isoline_interval}{TODO}

\item{color_scale}{TODO}

\item{color_map}{TODO}

\item{line_width}{TODO}

\item{line_color}{TODO}

\item{label_size}{TODO}

\item{label_color}{TODO}

\item{scale_units}{TODO}

\item{wrap_cohort_labels}{TODO}
}
\description{
TEXT
}
\examples{
# Hexamap for Overdose Deaths among White Men in the United States from 1999 through 2018
data.df <- read.csv("overdose_deaths_white_men_1999_2018.csv") #data can be downloaded from CDC Wonder.
data <- as.matrix(data.df) # convert to a matrix and exclude the row names
create_hexamap(data = data,  #matrix: age as rows, period as columns
               first_age = 15,
               first_period = 1999,
               interval = 1, 
               first_age_isoline = 20,
               first_period_isoline = 2000,
               isoline_interval = 10, 
               color_scale = c(0,40))

}
