% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_APChexamap.R
\name{plot_APChexamap}
\alias{plot_APChexamap}
\title{Hexamap of an APC surface}
\usage{
plot_APChexamap(
  data,
  first_age,
  first_period,
  interval,
  first_age_isoline = NULL,
  first_period_isoline = NULL,
  isoline_interval = NULL,
  color_scale = NULL,
  color_map = NULL,
  line_width = 0.5,
  line_color = "grey",
  label_size = 0.5,
  label_color = "black",
  scale_units = "Rate",
  wrap_cohort_labels = TRUE
)
}
\arguments{
\item{data}{TODO}

\item{first_age}{TODO}

\item{first_period}{TODO}

\item{interval}{TODO}

\item{first_age_isoline}{TODO}

\item{first_period_isoline}{TODO}

\item{isoline_interval}{TODO}

\item{color_scale}{TODO}

\item{color_map}{TODO}

\item{line_width}{TODO}

\item{line_color}{TODO}

\item{label_size}{TODO}

\item{label_color}{TODO}

\item{scale_units}{TODO}

\item{wrap_cohort_labels}{TODO}
}
\description{
TODO in analogy to \code{\link{plot_APCheatmap}}, allow to plot an observed
mean structure of a metric variable, or alternatively an estimated smoothed
mean structure based on a GAM model.
}
\examples{
# TODO

}
\seealso{
\code{\link{plot_APCheatmap}}
}
\author{
Hawre Jalal \email{hjalal@pitt.edu},
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
