% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density_categorical}
\alias{plot_density_categorical}
\title{Internal helper to plot a categorical density}
\usage{
plot_density_categorical(
  dat,
  y_var,
  dat_highlightDiagonals = NULL,
  weights_var = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{dat}{Dataset with columns \code{period} and \code{age} and the
main variable specified through argument \code{y_var}.}

\item{y_var}{Character name of the main variable to be plotted.}

\item{dat_highlightDiagonals}{Optional dataset created by
\code{\link{create_highlightDiagonalData}} to highlight specific diagonals
in a density matrix.}

\item{weights_var}{Optional character name of a weights variable used to
project the results in the sample to some population.}

\item{xlab, ylab}{Optional plot annotations.}
}
\description{
Internal helper function to plot one categorical density, to be called from
within \code{\link{plot_density}}.
}
