% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointMarginalAPCEffects.R
\name{plot_jointMarginalAPCEffects}
\alias{plot_jointMarginalAPCEffects}
\title{Joint plot to compare the marginal APC effects of multiple models}
\usage{
plot_jointMarginalAPCEffects(model_list, dat)
}
\arguments{
\item{model_list}{A list of regression models estimated with
\code{\link[mgcv]{gam}}. If the list is named, the names are used as labels.}

\item{dat}{Dataset as passed to \code{gam} for model estimation, containing
columns \code{period} and \code{age}.}
}
\description{
This function creates a joint plot of the marginal APC effects of multiple
estimated models. It creates a plot with one pane per age, period and
cohort effect, each containing one lines for each estimated model.
}
\details{
If the model was estimated with a log or logit link, the function
automatically performs an exponential transformation of the effect.
}
