% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jointMarginalAPCeffects.R
\name{plot_jointMarginalAPCeffects}
\alias{plot_jointMarginalAPCeffects}
\title{Joint plot to compare the marginal APC effects of multiple models}
\usage{
plot_jointMarginalAPCeffects(model_list, dat, vlines_list = NULL, ylab = NULL)
}
\arguments{
\item{model_list}{A list of regression models estimated with
\code{\link[mgcv]{gam}}. If the list is named, the names are used as labels.}

\item{dat}{Dataset as passed to \code{gam} for model estimation, containing
columns \code{period} and \code{age}.}

\item{vlines_list}{Optional list that can be used to highlight the borders of
specific age groups, time intervals or cohorts. Each element must be a
numeric vector of values on the x-axis where vertical lines should be drawn.
The list can maximally have three elements and must have names out of
\code{c("age","period","cohort"}.}

\item{ylab}{Optional title for the y-axis.}
}
\description{
This function creates a joint plot of the marginal APC effects of multiple
estimated models. It creates a plot with one pane per age, period and
cohort effect, each containing one lines for each estimated model.
}
\details{
If the model was estimated with a log or logit link, the function
automatically performs an exponential transformation of the effect.
}
\examples{
library(APCtools)
library(mgcv)

data(travel)
model_pure <- gam(mainTrip_distance ~ te(age, period), data = travel)
model_cov  <- gam(mainTrip_distance ~ te(age, period) + s(household_income),
                  data = travel)

model_list <- list("pure model"      = model_pure,
                   "covariate model" = model_cov)

plot_jointMarginalAPCeffects(model_list, dat = travel)

# mark specific cohorts
plot_jointMarginalAPCeffects(model_list, dat = travel,
                             vlines_list = list("cohort" = c(1966.5,1982.5,1994.5)))

}
