% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_variable}
\alias{plot_variable}
\title{Distribution plot of one variable against one APC dimension}
\usage{
plot_variable(
  dat,
  variable,
  apc_dimension = "period",
  log_scale = FALSE,
  geomBar_position = "fill"
)
}
\arguments{
\item{dat}{Dataset containing columns \code{age}, \code{period} and
\code{cohort}.}

\item{variable}{Character name of the variable to plot.}

\item{apc_dimension}{One of \code{c("age","period","cohort")}. Defaults to
\code{"period"}.}

\item{log_scale}{Indicator if the visualized variable should be log10
transformed. Only used if the variable is numeric. Defaults to FALSE.}

\item{geomBar_position}{Value passed to \code{\link[ggplot2]{geom_bar}} as
\code{position} argument. Only used if the visualized variable is categorical.
Defaults to \code{"fill"}.}
}
\description{
Plot the distribution of one variable in the data against age, period or
cohort. Creates a stacked bar plot for categorical variables and boxplots
for metric variables.
}
\examples{
library(APCtools)
data(travel)

# plot a metric variable
plot_variable(dat = travel, variable = "mainTrip_distance",
              apc_dimension = "period", log_scale = TRUE)

# plot a categorical variable
plot_variable(dat = travel, variable = "household_size", apc_dimension = "period")
plot_variable(dat = travel, variable = "household_size", apc_dimension = "period",
              geomBar_position = "stack")

}
\author{
Alexander Bauer \email{alexander.bauer@stat.uni-muenchen.de}
}
